+++ junos/provider.go
@@ -230,6 +230,7 @@ func Provider() *schema.Provider {
 			"junos_interface_logical_info":      dataSourceInterfaceLogicalInfo(),
 			"junos_interface_physical":          dataSourceInterfacePhysical(),
 			"junos_interfaces_physical_present": dataSourceInterfacesPhysicalPresent(),
+			"junos_routes":                      dataSourceRoutes(),
 			"junos_routing_instance":            dataSourceRoutingInstance(),
 			"junos_security_zone":               dataSourceSecurityZone(),
 			"junos_system_information":          dataSourceSystemInformation(),